+++ test/controllers/kwargs_test.rb
@@ -2,12 +2,12 @@ require 'test_helper'
 
 class KwargsControllerTest < ActionController::TestCase
   def setup
-    Controller::Testing::Kwargs.raise_exception
+    Rails::ForwardCompatibleControllerTests.raise_exception
   end
 
   %w[get post patch put head delete].each do |verb|
     define_method("test_#{verb}_old_params_only") do
-      Controller::Testing::Kwargs.ignore
+      Rails::ForwardCompatibleControllerTests.ignore
 
       send(verb.to_sym, :test_kwargs, hello: 'world')
       assert_equal({ 'hello' => 'world' }, assigns(:params))
@@ -16,7 +16,7 @@ class KwargsControllerTest < ActionController::TestCase
     end
 
     define_method("test_#{verb}_old_params_only__outputs_deprecation") do
-      Controller::Testing::Kwargs.deprecate
+      Rails::ForwardCompatibleControllerTests.deprecate
 
       assert_deprecated do
         send(verb.to_sym, :test_kwargs, hello: 'world')
@@ -41,7 +41,7 @@ class KwargsControllerTest < ActionController::TestCase
     end
 
     define_method("test_xhr_#{verb}_old_params_only") do
-      Controller::Testing::Kwargs.ignore
+      Rails::ForwardCompatibleControllerTests.ignore
 
       xhr verb.to_sym, :test_kwargs, hello: 'world'
       assert_equal({ 'hello' => 'world' }, assigns(:params))
@@ -50,7 +50,7 @@ class KwargsControllerTest < ActionController::TestCase
     end
 
     define_method("test_xhr_#{verb}_old_params_only__outputs_deprecation") do
-      Controller::Testing::Kwargs.deprecate
+      Rails::ForwardCompatibleControllerTests.deprecate
 
       assert_deprecated do
         xhr verb.to_sym, :test_kwargs, hello: 'world'
@@ -75,7 +75,7 @@ class KwargsControllerTest < ActionController::TestCase
     end
 
     define_method("test_#{verb}_old_headers_only") do
-      Controller::Testing::Kwargs.ignore
+      Rails::ForwardCompatibleControllerTests.ignore
 
       send(verb.to_sym, :test_kwargs, nil, 'Hello': 'world')
       assert_equal('world', assigns(:hello_header))
@@ -84,7 +84,7 @@ class KwargsControllerTest < ActionController::TestCase
     end
 
     define_method("test_#{verb}_old_headers_only__outputs_deprecation") do
-      Controller::Testing::Kwargs.deprecate
+      Rails::ForwardCompatibleControllerTests.deprecate
 
       assert_deprecated do
         send(verb.to_sym, :test_kwargs, nil, 'Hello': 'world')
@@ -108,7 +108,7 @@ class KwargsControllerTest < ActionController::TestCase
     end
 
     define_method("test_xhr_#{verb}_old_headers_only") do
-      Controller::Testing::Kwargs.ignore
+      Rails::ForwardCompatibleControllerTests.ignore
 
       xhr(verb.to_sym, :test_kwargs, nil, 'Hello': 'world')
       assert_equal('world', assigns(:hello_header))
@@ -117,7 +117,7 @@ class KwargsControllerTest < ActionController::TestCase
     end
 
     define_method("test_xhr_#{verb}_old_headers_only__outputs_deprecation") do
-      Controller::Testing::Kwargs.deprecate
+      Rails::ForwardCompatibleControllerTests.deprecate
 
       assert_deprecated do
         xhr(verb.to_sym, :test_kwargs, nil, 'Hello': 'world')
@@ -141,7 +141,7 @@ class KwargsControllerTest < ActionController::TestCase
     end
 
     define_method("test_#{verb}_old_params_and_headers") do
-      Controller::Testing::Kwargs.ignore
+      Rails::ForwardCompatibleControllerTests.ignore
 
       send(verb.to_sym, :test_kwargs, { hello: 'world' }, { 'Hello': 'world' })
       assert_equal({ 'hello' => 'world' }, assigns(:params))
@@ -150,7 +150,7 @@ class KwargsControllerTest < ActionController::TestCase
     end
 
     define_method("test_#{verb}_old_params_and_headers__outputs_deprecation") do
-      Controller::Testing::Kwargs.deprecate
+      Rails::ForwardCompatibleControllerTests.deprecate
 
       assert_deprecated do
         send(verb.to_sym, :test_kwargs, { hello: 'world' }, { 'Hello': 'world' })
@@ -174,7 +174,7 @@ class KwargsControllerTest < ActionController::TestCase
     end
 
     define_method("test_xhr_#{verb}_old_params_and_headers") do
-      Controller::Testing::Kwargs.ignore
+      Rails::ForwardCompatibleControllerTests.ignore
 
       xhr verb.to_sym, :test_kwargs, { hello: 'world' }, { 'Hello': 'world' }
       assert_equal({ 'hello' => 'world' }, assigns(:params))
@@ -183,7 +183,7 @@ class KwargsControllerTest < ActionController::TestCase
     end
 
     define_method("test_xhr_#{verb}_old_params_and_headers__deprecated") do
-      Controller::Testing::Kwargs.deprecate
+      Rails::ForwardCompatibleControllerTests.deprecate
 
       assert_deprecated do
         xhr verb.to_sym, :test_kwargs, { hello: 'world' }, { 'Hello': 'world' }
@@ -207,7 +207,7 @@ class KwargsControllerTest < ActionController::TestCase
     end
 
     define_method("test_xhr_#{verb}_new_params_and_headers__does_not_output_warning") do
-      Controller::Testing::Kwargs.deprecate
+      Rails::ForwardCompatibleControllerTests.deprecate
 
       assert_not_deprecated do
         send(verb.to_sym, :test_kwargs, xhr: true, params: { hello: 'world' }, headers: { 'Hello': 'world' })