+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Rename gem from Controller::Testing::Kwargs to Rails::ForwardCompatibleControllerTests
+
 ## [1.0.3] - 2017-08-09
 
 ### Fixed